+++ src/main/java/umc/TripPiece/repository/CountryRepository.java
@@ -6,6 +6,6 @@ import umc.TripPiece.domain.Country;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface CountryRepository extends JpaRepository<Country, BigInteger> {
+public interface CountryRepository extends JpaRepository<Country, Long> {
     List<Country> findByNameContainingIgnoreCase(String name);
 }
